refactor: shared keybindings via message passing
Application events are widget/component agnostic, which means that it is
the responsibility of the currently active/focused widget/component to
react to a particular event. Widget agnostic events offer better
semantics for creating a shared keybinding system.

As a result, some keybindings will be context sensitive. For example the
keybinding "q" closes nested/expanded views but also exits the
application if no views are open. To support this behavior, a message
passing system is used to allow bidirectional communication between the
top level application and widgets. Therefore, if a widget has no
currently active view and receives a "q" key event, it will send a
message to upstream to close the application via message passing.