fix: expanded item changes when feed is updated
This is due to tracking the index of the item that is currently
expanded. Since the feed is asynchronously updated (optimisticly render
the moment there is something to show), the index of an item may change
as we reorder the items in the feed upon receiving the update.

As a result, to prevent the expanded item from being lost, an unique
identifer is now used to track the item that is expanded.

Additional refactors:

- Instead of giving items a fallback title/url if they are missing, we
  track the existence using `Option` and render the item differently