perf: cache rendered content of expanded feed item
Some feed items contain content that is rather long and hence expensive
to render. The rendering load is mostly caused by dynamic word wrapping
to fit the content into the available terminal width. The longer the
content, based on the number of lines, the more expensive it is to
render since dynamic word wrapping is done at each render cycle.

This commit introduces a cache for the rendered content of expanded feed
to prevent the need to repeatedly wrap the content when the terminal
width has not changed since the last render.

Although the cache brought a significant performance improvement in
terms of sustained FPS, there is still a noticeable dip in FPS when
rendering feed item with long content. My guess is that this is due to
how the third party `ScrollView` widget is implemented. Without wrapping
the content in a `ScrollView`, there is no noticeable dip in FPS.
Perhaps, we need to implement our own scrollbar while leverage the
native `Paragraph` widget scrolling functions for rendering the content.