perf: custom scrollbar and scroll offset handling for feed item content
Instead of relying on `tui-scrollview` for scrolling capabilities, I am
rolling my own scrollbar to ensure that we get native (`ratatui`)
performance as much as possible. With the new scrollbar, there is no
sharp FPS dip even when rendering long (~500+ lines) content.

However, with the new scrollbar, another interesting performance issue
was discovered, relating to `ratatui::Paragraph::scroll`. The sustainted
FPS when rendering a long content with no scroll offset (aka
`scroll((0,0))`) was significantly faster than when rendering with
increasingly larger scroll offsets. FPS drops proportionally to the
magnitude of the scroll offset (vertically). 

The implementation of `scroll` seems to contain a loop that goes through
each line of the content based on the scroll offset. When no offset is
applied, it does not iterate through any lines. I am still not fully
sure what the underlying reason is but for my use case, I didn't observe
any rendering issues when I implemented my own scroll offset
implementation.

Instead of utilizing `ratatui::Paragraph::scroll`, I pass in a slice of
the content using `content[scroll_offset..]` to the `Paragraph` widget
and render it directly. With this change, the FPS remains stable no
matter where the scroll offset is set. The only downside is that I would
need to manually synchronize the scroll offset such that it does not go
beyond the content length (leading to an invalid slices).