fun: tty event debugger script
Makes use of `crossterm` as the backend for parsing and event streams,
as that is the backed used by `rssterm`.

This script is intended to replicate and get more information about the
scroll event flooding bug that was discovered when building `rssterm`.
Turns out someone also had the same issue reported in the `crossterm`
repo: https://github.com/crossterm-rs/crossterm/issues/854