feat: activity/loading indicator when pulling feeds
The render rate (step rate) of the throbber should be independent from
the application's FPS, so that it does not vary based on user's choice
of FPS. If the FPS is too high, the throbber will be too fast and it no
longer looks like a throbber. As such, a wrapper around the throbber
widget is added to control a fixed step rate (250ms).